% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeDistance2Coast.R
\name{ComputeDistance2Coast}
\alias{ComputeDistance2Coast}
\title{Compute distance to coast}
\usage{
ComputeDistance2Coast(location, mapdir)
}
\arguments{
\item{location}{vector of locations (longitude,latitude)}

\item{mapdir}{directory containing shape file with england coastline}
}
\value{
distance  distance to coast and coordinates of nearest point
}
\description{
Compute distance to coast
}
\examples{
dist <- ComputeDistance2Coast(location,mapdir)
}
