% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeNICDistance.R
\name{ComputeNICDistance}
\alias{ComputeNICDistance}
\title{ComputeNICDistance(newdata,dest)
Compute distances from each observation in newdata to each destination in dest.}
\usage{
ComputeNICDistance(newdata, dest)
}
\arguments{
\item{newdata}{dataframe containing (latitude,longitude)}

\item{dest}{dataframe containing [dest_name,dest_name_short,latitude,longitude]}
}
\value{
newdata dataframe containing results
}
\description{
ComputeNICDistance(newdata,dest)
Compute distances from each observation in newdata to each destination in dest.
}
\examples{
newdata<-ComputeNICDistance(olddata,dest)
}
\keyword{distance}
\keyword{geographic}
