% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRank.R
\name{CreateRank}
\alias{CreateRank}
\title{Rank each variable of type travelmode.}
\usage{
CreateRank(m2data, travelmode)
}
\arguments{
\item{m2data}{dataframe, with columns of travel times}

\item{travelmode}{c("drive","trans")}
}
\value{
m2data
}
\description{
Rank each variable of type travelmode.
}
\examples{
Finds all columns with travel times of current type, and ranks the travel times
from lowest to highest. For example, if travelmode=="drive", it finds all columns
containing drive_times and ranks them. Creates new variables in dataframe
containing the rankings.#'
m2data<-CreateRank(m2data,"drive")
m2data<-CreateRank(m2data,"trans")
}
