% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNearestAONB.R
\name{FindNearestAONB}
\alias{FindNearestAONB}
\title{Find nearest point on AONB to houses and compute travel time}
\usage{
FindNearestAONB(houses, regino_polygon, region_id, dirs)
}
\arguments{
\item{houses}{(n x 2) matrix of coordinates (longitude,latitude)}

\item{region_id}{index for region}

\item{dirs}{list of directories: (dirs$mapdir,dirs$traveldir)}

\item{region_polygon}{region polygon (bounding box or convex hull)}
}
\value{
newdata   dataframe with (AONB_lon,AONB_lat,distance_AONB,drive_AONB)
}
\description{
Find nearest point on AONB to houses and compute travel time
}
\examples{
houses    <- as.matrix(m2data[,c("longitude","latitude")])
region_id <- 1
dirs$mapdir <- paste0(RootDir,"/data/maps")
dirs$traveldir <- paste0(RootDir,"/code/output/TravelTime")
newdata<-FindNearestAONB(houses,region_polygon,region_id,dirs)
}
