% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNearestCoastPoint.R
\name{FindNearestCoastPoint}
\alias{FindNearestCoastPoint}
\title{Find nearest point on coast to houses and compute travel time}
\usage{
FindNearestCoastPoint(houses, region_polygon, region_id, dirs, basis = "glp")
}
\arguments{
\item{houses}{(n x 2) matrix of coordinates (longitude,latitude)}

\item{region_polygon}{polygon for region (bounding box or convex hull)}

\item{region_id}{index for region}

\item{dirs}{list of directories: (dirs$mapdir,dirs$traveldir)}

\item{basis}{c("glp","tensor","cheb") basis for travel time model}
}
\value{
newdata   dataframe with (coast_lon,coast_lat,distance_coast,drive_coast)
}
\description{
Find nearest point on coast to houses and compute travel time
}
\examples{
houses    <- as.matrix(m2data[,c("longitude","latitude")])
region_id <- 1
dirs$mapdir <- paste0(RootDir,"/data/maps")
dirs$traveldir <- paste0(RootDir,"/code/output/TravelTime")
newdata<-FindNearestCoastPoint(houses,region_polygon,region_id,dirs)
}
