% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNearestRoad.R
\name{FindNearestRoad}
\alias{FindNearestRoad}
\title{Find nearest road to houses and compute travel time}
\usage{
FindNearestRoad(houses, region_polygon, roadtype = "motorways", region_id,
  dirs, basis = "glp")
}
\arguments{
\item{houses}{(n x 2) matrix of coordinates (longitude,latitude)}

\item{region_polygon}{polygon for region (bounding box or convexhull)}

\item{roadtype}{either "motorways" or "aroads"}

\item{region_id}{index for region}

\item{dirs}{list of directories: (dirs$roaddir,dirs$traveldir)}

\item{basis}{c("glp","tensor","cheb") basis for travel time model}
}
\value{
newdata   dataframe with (roadtype_lon,roadtype_lat,distance_roadtype,drive_roadtype)
}
\description{
Find nearest road to houses and compute travel time
}
\examples{
houses    <- as.matrix(m2data[,c("longitude","latitude")])
region_id <- 1
dirs$traveldir <- paste0(RootDir,"/code/output/TravelDir")
dirs$roaddir <- paste0(RootDir,"/data/roads")
newdata<-FindNearestRoad(houses,roadtype="motorways",region_id,dirs)
}
