% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNearestStation.R
\name{FindNearestStation}
\alias{FindNearestStation}
\title{Find nearest station to houses and compute travel time}
\usage{
FindNearestStation(houses, region_id, RootDir, TravelDir, basis = "glp",
  stationtype = "rail")
}
\arguments{
\item{houses}{(n x 2) matrix of coordinates (longitude,latitude)}

\item{region_id}{index for region}

\item{RootDir}{directory name of root directory}

\item{TravelDir}{directory name of travel model directory}

\item{basis}{c("glp","tensor","cheb") basis for travel time model}
}
\value{
newdata   dataframe with (station_lon,station_lat,distance_station,drive_station)
}
\description{
Find nearest station to houses and compute travel time
}
\examples{
houses    <- as.matrix(m2data[,c("longitude","latitude")])
region_id <- 1
RootDir   <- "/Users/hedonic/NIC"
TravelDir <- paste0(RootDir,"/output/TravelDir")
newdata<-FindNearestStation(houses,region_id,RootDir,TravelDir)
}
