% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTownNames.R
\name{GetTownNames}
\alias{GetTownNames}
\title{Lookup coordinates and town names for each postcode}
\usage{
GetTownNames(m1data)
}
\arguments{
\item{m1data}{}
}
\value{
m1data
}
\description{
Lookup coordinates and town names for each postcode
}
\examples{
m1data<-GetTownNames(m1data)
}
