% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap1.R
\name{heatmap1}
\alias{heatmap1}
\title{heatmap1(x,y,z,filter,resolution,zlabel,outfile)
This function plots a heatmap for z = f(x,y)}
\usage{
heatmap1(x, y, z, filter, resolution, zlabel, outfile, places, route,
  routelabel)
}
\arguments{
\item{x}{vector of \code{x} coordinates}

\item{y}{vector of \code{y} coordinates}

\item{z}{vector of \code{z} coordinates}

\item{filter}{logical to subset \code{(x,y,z)}}

\item{zlabel}{string   label for \code{z}}

\item{outfile}{name of output file}

\item{places}{n x 3 matrix of places, (name,lon,lat)}

\item{route}{n x 3 matrix of route, (name,lon,lat)}

\item{routelabel}{"road" or "rail"}
}
\value{
plot1   details of plot
}
\description{
heatmap1(x,y,z,filter,resolution,zlabel,outfile)
This function plots a heatmap for z = f(x,y)
}
\examples{
h<-heatmap1(x,y,z,filter=isample,resolution=0.005,zlabel="z",outfile="h1.eps")
}
\keyword{heatmap}
