% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTravelTime.R
\name{predictTravelTime}
\alias{predictTravelTime}
\title{predictTravelTime(origin,dest,region_id,TravelDir,travelmode,drop_outlier)
This function predicts travel times from orgin to dest, for region_id}
\usage{
predictTravelTime(origin, dest, region_id, TravelDir, travelmode = "drive",
  drop_outlier = TRUE)
}
\arguments{
\item{origin}{data on (longitude,latitude) for origins}

\item{dest}{destinations with (longitude,latitude)}

\item{region_id}{id number for region between 1 and 11}

\item{TravelDir}{directory containing travel model results}

\item{drop_outlier}{TRUE to drop top 0.001 outliers}

\item{travelmodel}{c("drive","trans")}
}
\value{
origin         Add predicted travel times to origin
}
\description{
predictTravelTime(origin,dest,region_id,TravelDir,travelmode,drop_outlier)
This function predicts travel times from orgin to dest, for region_id
}
\examples{
newdata <-predictTravelTime(origin,dest,region_id,TravelDir,travelmode="drive",drop_outlier=TRUE)
newdata <-predictTravelTime(origin,dest,region_id,TravelDir,travelmode="trans",drop_outlier=FALSE)
}
\keyword{time}
\keyword{travel}
