% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictTravelTime2.R
\name{predictTravelTime2}
\alias{predictTravelTime2}
\title{predictTravelTime2(orig_dest,region_id,TravelDir)
This function predicts travel times from orgin to dest, for region_id}
\usage{
predictTravelTime2(orig_dest, destname, region_id, TravelDir,
  printstr = "default", basis = "glp")
}
\arguments{
\item{orig_dest}{data on (longitude,latitude) for (origin,destination) pairs}

\item{destname}{short name of destination}

\item{region_id}{id number for region between 1 and 11}

\item{TravelDir}{directory containing travel model results}

\item{printstr}{name of destination to print to std. io.}

\item{basis}{c("glp","tensor","cheb"): corresponding travelmodel results must exist}
}
\value{
traveltime     Add predicted travel times to origin
}
\description{
predictTravelTime2(orig_dest,region_id,TravelDir)
This function predicts travel times from orgin to dest, for region_id
}
\examples{
printstr<-"default"
traveltime <-predictTravelTime2(orig_dest,destname,region_id,TravelDir,printstr,"glp")
printstr<-"Dartmoor"
traveltime <-predictTravelTime2(orig_dest,destname,region_id,TravelDir,printstr,"glp")
traveltime <-predictTravelTime2(orig_dest,destname,region_id,TravelDir,printstr,"cheb")
}
\keyword{time}
\keyword{travel}
